added load skeleton
